Data types

PauliString

PauliPropagation.PauliString — Type

PauliString(nqubits::Int, pauli::Symbol, qind::Integer, coeff=1.0)
PauliString(nqubits::Int, paulis::Vector{Symbol}, qinds::Vector{Integer}, coeff=1.0)

Constructor for a PauliString on nqubits qubits from a Symbol (:I, :X, :Y, :Z) or Vector{Symbol}. Provide the index or indices for those symbols as qind or qinds. The coefficient of the Pauli string in the Pauli sum defaults to 1.0.

PauliPropagation.PauliString — Type

PauliString(nqubits::Int, term::TermType, coeff::CoeffType)

PauliString is a struct that represents a Pauli string on nqubits qubits. Commonly term is an unsigned Integer. See the other PauliString constructors for higher-level usage.

Base.:* — Method

*(pstr::PauliString, c::Number)

Multiply a PauliString by a scalar c. Returns a new PauliString.

Base.:* — Method

*(pstr1::PauliString, pstr2::PauliString)

Perform a Pauli product of two PauliStrings.

Base.:/ — Method

/(pstr::PauliString, c::Number)

Divide a PauliString by a scalar c. Returns a new PauliString.

PauliPropagation.PropagationBase.coefftype — Method

coefftype(pstr::PauliString)

Get the coefficient type of a PauliString.

PauliPropagation.PropagationBase.numcoefftype — Method

numcoefftype(pstr::PauliString)

Get the type of the numerical coefficient of a PauliString. Will return the type of the output of numcoefftype(coefftype(pstr)).

PauliPropagation.nqubits — Method

nqubits(pstr::PauliString)

Get the number of qubits that the PauliString is defined on.

PauliPropagation.paulitype — Method

paulitype(pstr::PauliString)

Get the Pauli integer type of a PauliString.

PauliSum

PauliPropagation.PauliSum — Type

PauliSum(nqubits::Int, terms::Dict) <: AbstractPauliSum

PauliSum is a struct that represents a sum of Pauli strings acting on nqubits qubits. It is a wrapper around a dictionary Dict(Pauli string => coefficient}, where the Pauli strings are typically unsigned Integers for efficiency reasons.

PauliPropagation.PauliSum — Method

PauliSum(nqubits::Integer)

Contructor for an empty PauliSum on nqubits qubits. Element type defaults for Float64.

PauliPropagation.PauliSum — Method

PauliSum(CoeffType, nq::Int)

Contructor for an empty PauliSum on nqubits qubits. The type of the coefficients can be provided.

Base.copy — Method

Copy a PauliSum by copying its terms field.

PauliPropagation.nqubits — Method

nqubits(psum::PauliSum)

Get the number of qubits that the PauliSum is defined on.

VectorPauliSum

PauliPropagation.VectorPauliSum — Type

VectorPauliSum{TV,CV} <: AbstractPauliSum

VectorPauliSum is a struct that represents a sum of Pauli strings acting on nqubits qubits. It is a wrapper around two vectors: one for the Pauli strings (as unsigned Integers for efficiency reasons), and one for the coefficients. Using it defaults to multi-threaded operations where possible.

PauliPropagation.VectorPauliSum — Method

VectorPauliSum(nqubits::Int)

Constructor for an empty VectorPauliSum on nqubits qubits. Element type defaults for Float64.

PauliPropagation.VectorPauliSum — Method

VectorPauliSum(::Type{CT}, nqubits::Int)

Contructor for an empty VectorPauliSum on nqubits qubits. The type of the coefficients can be provided.

PauliPropagation.nqubits — Method

nqubits(vpsum::VectorPauliSum)

Get the number of qubits that the VectorPauliSum is defined on.

Conversions