Propagation

Generics

PauliPropagation.PropagationBase.propagate — Function

propagate(circ, pstr::PauliString, thetas=nothing; min_abs_coeff=eps(), max_weight=Inf, max_freq=Inf, max_sins=Inf, customtruncfunc=nothing, kwargs...)

Propagate a PauliString through the circuit circ in the Heisenberg picture. This means that the circuit is applied to the Pauli string in reverse order, and the action of each gate is its conjugate action. Parameters for the parametrized gates in circ are given by thetas, and need to be passed as if the circuit was applied as written in the Schrödinger picture. If thetas are not passed, the circuit must contain only non-parametrized StaticGates. Default truncations are min_abs_coeff, max_weight, max_freq, and max_sins. max_freq, and max_sins will lead to automatic conversion if the coefficients are not already wrapped in suitable PathProperties objects. A custom truncation function can be passed as customtruncfunc with the signature customtruncfunc(pstr::PauliStringType, coefficient)::Bool. Further kwargs are passed to the lower-level functions applymergetruncate!, applytoall!, and apply.

PauliPropagation.PropagationBase.propagate — Function

propagate(circ, psum::AbstractPauliSum, thetas=nothing; min_abs_coeff=eps(), max_weight=Inf, max_freq=Inf, max_sins=Inf, customtruncfunc=nothing, kwargs...)

Propagate a Pauli sum psum through the circuit circ in the Heisenberg picture. This means that the circuit is applied to the Pauli sum in reverse order, and the action of each gate is its conjugate action. The Pauli sum psum is deepcopied and passed into the in-place propagation function PropagationBase.propagate!(). Parameters for the parametrized gates in circ are given by thetas, and need to be passed as if the circuit was applied as written in the Schrödinger picture. If thetas are not passed, the circuit must contain only non-parametrized StaticGates. Default truncations are min_abs_coeff, max_weight, max_freq, and max_sins. max_freq, and max_sins will lead to automatic conversion if the coefficients are not already wrapped in suitable PathProperties objects. A custom truncation function can be passed as customtruncfunc with the signature customtruncfunc(pstr::PauliStringType, coefficient)::Bool. Further kwargs are passed to the lower-level functions applymergetruncate!, applytoall!, and apply.

PauliPropagation.PropagationBase.propagate! — Function

propagate!(circ, pstr::PauliString, thetas=nothing; min_abs_coeff=eps(), max_weight=Inf, max_freq=Inf, max_sins=Inf, customtruncfunc=nothing, kwargs...)

In-place version of propagate() for a PauliString. This is only a convenience function, because the PauliString is converted into a PauliSum internally. If max_freq, and max_sins are used without the coefficients already being wrapped in suitable PathProperties objects, an error is thrown.

PauliPropagation.PropagationBase.truncate! — Method

truncate!(psum::AbstractPauliSum; minabscoeff=eps(), maxweight=Inf, maxfreq=Inf, maxsins=Inf, customtruncfunc=nothing, kwargs...) truncate!(propcache::AbstractPauliPropagationCache; minabscoeff=eps(), maxweight=Inf, maxfreq=Inf, max_sins=Inf, customtruncfunc=nothing, kwargs...)

Truncation function for AbstractPauliPropagationCaches that combines multiple truncation criteria. The default truncation criteria are:

min_abs_coeff: Truncates Pauli strings with absolute coefficient below this value.

max_weight: Truncates Pauli strings with weight (number of non-identity Paulis) above this value.

max_freq: Truncates Pauli strings with frequency (number of cosine factors in coefficient) above this value.

max_sins: Truncates Pauli strings with number of sine factors in coefficient above this value.

A custom truncation function can be passed as customtruncfunc with the signature customtruncfunc(pstr, coeff)::Bool.

This function combines all truncation criteria into a single truncation function truncfunc() calls PropagationBase.truncate!(truncfunc, prop_cache).

Specializations

PauliPropagation.PropagationBase.apply — Method

apply(gate::TransferMapGate, pstr, coeff)

Apply a TransferMapGate to an integer Pauli string and its coefficient. The outcomes are determined by the transfer_map of the gate.

PauliPropagation.PropagationBase.applytoall! — Method

applytoall!(gate::AmplitudeDampingNoise, theta, psum, aux_psum; kwargs...)

Overload of applytoall! for AmplitudeDampingNoise gates. It fixes the type-instability of the apply() function and reduces moving Pauli strings between psum and auxpsum. psum and `auxpsum` are merged later.

PauliPropagation.PropagationBase.applytoall! — Method

applytoall!(gate::FrozenGate, thetas, psum, aux_psum; kwargs...)

Overload of applytoall! for FrozenGates. Re-directs to applytoall! for the wrapped FrozenGate.gate with the frozen parameter.

PauliPropagation.PropagationBase.applytoall! — Method

applytoall!(gate::PauliNoise, p, psum, aux_psum; kwargs...)

Overload of applytoall! for PauliNoise gates with noise strength p. It changes the coefficients in-place and does not require the aux_psum, which stays empty.

PauliPropagation.PropagationBase.applytoall! — Method

applytoall!(gate::PauliRotation, theta, psum, aux_psum; kwargs...)

Overload of applytoall! for PauliRotation gates. It fixes the type-instability of the apply() function and reduces moving Pauli strings between psum and aux_psum. psum and aux_psum are merged later.

PauliPropagation.PropagationBase.applytoall! — Method

applytoall!(gate::TGate, prop_cache::AbstractPauliPropagationCache; kwargs...)

Overload of applytoall!() for TGate(qind). Redirects to a PauliRotation(:Z, qind) with angle π/4.