State Overlap

Base.filter! — Method

filter!(filterfunc::Function, psum::PauliSum)

Filter a PauliSum in-place by removing all Pauli strings for which filterfunc(pstr, coeff) returns false.

Base.filter — Method

filter(filterfunc::Function, psum::PauliSum)

Return a filtered PauliSum by removing all Pauli strings for which filterfunc(pstr, coeff) returns false.

PauliPropagation.overlapbyorthogonality — Method

overlapbyorthogonality(orthogonalfunc::Function, psum::PauliSum)
overlapbyorthogonality(orthogonalfunc::Function, pstr::PauliString)
overlapbyorthogonality(orthogonalfunc::Function, pstr::Integer)

Overlap a PauliSum, PauliString or integer Pauli string with a state or operator via function that returns true if a Pauli string is orthogonal or not. If a Pauli string is orthogonal, it does not contribute, otherwise it contributes with its coefficient. This is particularly useful for overlaps with stabilizer states. An example orthogonalfunc is containsXorY which returns true if a Pauli string contains an X or Y Pauli.

PauliPropagation.overlapwithcomputational — Method

overlapwithcomputational(psum::PauliSum, onebitinds)
overlapwithcomputational(pstr::PauliString, onebitinds)

Calculates the overlap of a Pauli sum or Pauli string with the computational basis state which has one-bits at all specified indices and zero-bits elsewhere. If |x><x| is a computational basis state, it we compute Tr[psum * |x><x|] = <x|psum|x> or Tr[pstr * |x><x|] = <x|pstr|x>. For example, overlapwithcomputational(psum, [1,2,4]) returns the overlap with |1101000...>.

PauliPropagation.overlapwithmaxmixed — Method

overlapwithmaxmixed(psum::PauliSum)

Calculates the overlap of a PauliSum with the maximally mixed state I/2^n, i.e., Tr[psum * I/2^n].

PauliPropagation.overlapwithplus — Method

overlapwithplus(psum) 
overlapwithplus(pstr)

Calculates the overlap of a Pauli sum or Pauli string with the plus state |+><+|, i.e. Tr[psum * |+><+|] = <+|psum|+> or Tr[pstr * |+><+|] = <+|pstr|+>.

PauliPropagation.overlapwithzero — Method

overlapwithzero(psum) 
overlapwithzero(pstr)

Calculates the overlap of a Pauli sum with the zero state |0><0|, i.e., Tr[psum * |0><0|] = <0|psum|0> or Tr[pstr * |0><0|] = <0|pstr|0>.

PauliPropagation.plusfilter! — Method

zerofilter!(psum)

Filter a Pauli sum in-place with only Pauli strings that are not orthogonal to the plus state |+><+|.

PauliPropagation.plusfilter — Method

plusfilter(psum)

Return a filtered Pauli sum with only Pauli strings that are not orthogonal to the plus state |+><+|.

PauliPropagation.scalarproduct — Method

scalarproduct(psum1::PauliSum, psum2::PauliSum)
scalarproduct(pstr::PauliString, psum::PauliSum)
scalarproduct(psum::PauliSum, pstr::PauliString)
scalarproduct(pstr1::PauliString, pstr2::PauliString)

Calculates the scalar product between any combination of PauliSum and PauliString. This calculates the sum of the products of their coefficients for all Pauli strings that are present . Important: This is not equivalent to the trace Tr[psum1 * psum2] but instead Tr[psum1 * psum2]/2^n, and equivalently for Pauli strings.

PauliPropagation.zerofilter! — Method

zerofilter!(psum)

Filter a Pauli sum in-place with only Pauli strings that are not orthogonal to the zero state |0><0|.

PauliPropagation.zerofilter — Method

zerofilter(psum)

Return a filtered Pauli sum with only Pauli strings that are not orthogonal to the zero state |0><0|.